Fixed typos

+++ 01_Bare_metal/RPIOS/include/utills.h
@@ -4,4 +4,4 @@
 
 void delay(u64 ticks);
 void put32(u64 address,u32 value);
-u32 get32(u32 address);
+u32 get32(u64 address);
